services/worker-gpu/requirements.txt
torch>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
xformers>=0.0.23
opencv-python>=4.8.0
Pillow>=10.1.0
flask>=3.0.0
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0

